poetry-demo/pyproject.toml
[tool.poetry]
name = "poetry-demo"
version = "0.1.0"
description = ""
authors = ["amitaiturkel <amitai.turkel@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
tox = "^4.11.3"
pre-commit = "^3.4.0"
typing-extensions = "^4.8.0"
pytest = "^7.4.2"
nats-py = "^2.4.0"
argparse = "^1.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
